pyproject.toml
[project]
name = "toplinguagens"
version = "0.1.0"
description = ""
authors = [
    {name = "Bergsmf",email = "bergsmf@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "pandas (>=2.3.0,<3.0.0)",
    "pytest (>=8.4.0,<9.0.0)",
    "ruff (>=0.11.13,<0.12.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "taskipy (>=1.14.1,<2.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "streamlit (>=1.45.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
pre_run = 'poetry run python toplinguagens/loader.py'
run = 'poetry run streamlit run toplinguagens/app.py'
pre_test = 'task lint'
test = 'pytest -v'
[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

